+++ Samples~/Assets/Tests/Runtime/Scripts/XLog.Std.Test.cs
@@ -33,11 +33,11 @@ public class TestXLogStd
     [Test]
     public void Prefs()
     {
-        var panel = ScriptableObject.CreateInstance<XLog.StdAdapter.Prefs>();
+        var prefsEditor = new XLog.StdAdapter.Prefs() as XPrefs.IEditor;
 
         // Save
         var targetPrefs = new XPrefs.IBase();
-        panel.OnSave(source: new XPrefs.IBase(), target: targetPrefs);
+        prefsEditor.OnSave(source: new XPrefs.IBase(), target: targetPrefs);
 
         var targetConfig = targetPrefs.Get<XPrefs.IBase>(XLog.StdAdapter.Prefs.Config);
         Assert.NotNull(targetConfig, "Log/Std 配置项应当存在。");
@@ -46,7 +46,7 @@ public class TestXLogStd
         Assert.AreEqual(XLog.StdAdapter.Prefs.ColorDefault, targetConfig.GetBool(XLog.StdAdapter.Prefs.Color), "Log/Std/Color 配置项应当存在。");
 
         // Apply
-        panel.OnApply(source: targetPrefs, target: targetPrefs, asset: false, remote: true);
+        prefsEditor.OnApply(source: targetPrefs, target: targetPrefs, asset: false, remote: true);
         targetConfig = targetPrefs.Get<XPrefs.IBase>(XLog.StdAdapter.Prefs.Config);
         Assert.IsNull(targetConfig, "Log/Std 配置项在远端应当被移除。");
     }